Block add_index without concurrently algorithm

+++ lib/safe_migrations/migration_ext.rb
@@ -25,7 +25,7 @@ module SafeMigrations
       end
 
       def method_missing_with_safety(method, *args, &block)
-        if safe? || bypassing_safety_checks? || !UNSAFE_METHODS.include?(method)
+        if safe? || bypassing_safety_checks?
           return method_missing_without_safety(method, *args, &block)
         end
 
@@ -34,7 +34,14 @@ module SafeMigrations
           raise UnsafeRemoveColumn
         when :drop_table
           raise UnsafeDropTable
+        when :add_index
+          options = args[2]
+          unless (options && options[:algorithm] == :concurrently)
+            raise UnsafeAddIndex
+          end
         end
+
+        method_missing_without_safety(method, *args, &block)
       end
     end
